// src/components/form-components/MySafetyCSV/MySafetyCSV.types.ts
export interface MySafetyCSVData {
  actualValue: number;
  createdByEmail?: string;
  createdByName: string;
  createdByTelephone?: string;
  crimeReferenceNumber?: string;
  crimeType: string[];
  dateCreated: Date;
  dateOccurred: Date;
  description: string;
  emergencyServiceNumber?: string;
  emergencyServicesAttend?: boolean;
  emergencyServicesRef?: string;
  estimatedValue: number;
  incidentID: string;
  injuryAction?: string;
  officerName?: string;
  officerSerial?: string;
  officerTelephone?: string;
  site: string;
  specificArea: string;
  suspectApproxAge?: string;
  suspectsAddressLine1?: string;
  suspectsAddressLine2?: string;
  suspectsAnyScars?: string;
  suspectsBuild?: string;
  suspectsClothing?: string;
  suspectsCounty?: string;
  suspectsEthnicGroup?: string;
  suspectsEyeColour?: string;
  suspectsFacialHair?: string;
  suspectsFullName?: string;
  suspectsGender?: string;
  suspectsGlasses?: string;
  suspectsHairColour?: string;
  suspectsHeightImperial?: string;
  suspectsHeightMetric?: string;
  suspectsPiercings?: string;
  suspectsPostcode?: string;
  suspectsTattooDescription?: string;
  suspectsTattoos?: string;
  suspectsTelephone?: string;
  suspectsTitle?: string;
  suspectsTown?: string;
  suspectsWeightImperial?: string;
  suspectsWeightMetric?: string;
  timeOccurred: Date;
  vehicleColour?: string;
  vehicleColourOther?: string;
  vehicleMake?: string;
  vehicleRegistration?: string;
  vehicleType?: string;
  weaponsUsed?: string;
  weaponsUsedOther?: string;
  wereWeaponsUsed?: boolean;
}